requirements.txt
numpy
pandas
seaborn
matplotlib
pickle-mixin
scikit-learn
flask
nltk
tweepy
python-dotenv
vaderSentiment
flair